+++ src/components/common/Modal.tsx
@@ -18,7 +18,7 @@ const StyledPanel = styled.div`
   display: grid;
   background: var(--bg-color);
   box-shadow: 0 3px 8px rgb(0 0 0 / 24%);
-  padding: 4px;
+  padding: 12px;
   border-radius: 8px;
   white-space: nowrap;
   position: relative;